requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
pydantic==2.4.2
nltk==3.8.1
scikit-learn==1.3.2
